moves all routers to one folder
