+++ routes/recommendations-router.js
@@ -3,7 +3,7 @@ const axios = require("axios");
 const { authToken } = require("../auth/authenticate-middleware");
 const router = express.Router();
 
-const Recommendations = require("./recommendations-model");
+const Recommendations = require("../models/recommendations-model.js");
 
 /**
  * @api {get} /users/:user_id/recommendations
@@ -75,16 +75,19 @@ const Recommendations = require("./recommendations-model");
  */
 router.get("/:id/recommendations", (req, res) => {
   const { id } = req.params;
-  Recommendations.getLatestRecommendations(id)
-    .then((res) => {
-      console.log(res)
-      if(res){
-        const recParams = res
-      }
-    })
-    .catch((err) => {
-      console.log(err);
-    });
+  // * For use at a later time *
+  //
+  // Recommendations.getLatestRecommendations(id)
+  //   .then((res) => {
+  //     console.log(res)
+  //     if(res){
+  //       const recParams = res
+  //     }
+  //   })
+  //   .catch((err) => {
+  //     console.log(err);
+  //   });
+
   axios
     .post(
       process.env.RECOMMENDATION_URL,
